Extract InputContainer
