+++ src/Page.jsx
@@ -1,19 +1,13 @@
 import React from 'react';
 
-import Input from './Input';
 import ListContainer from './ListContainer';
+import InputContainer from './InputContainer';
 
-export default function Page({
-  taskTitle, onChangeTitle, onClickAddTask,
-}) {
+export default function Page() {
   return (
     <div>
       <h1>To-do</h1>
-      <Input
-        value={taskTitle}
-        onChange={onChangeTitle}
-        onClick={onClickAddTask}
-      />
+      <InputContainer />
       <ListContainer />
     </div>
   );
